+++ src/models/rule.ts
